Use a FolderSelector instance rather than perform inline
In order to make better injectable behaviours, introducing a class that can contain the "select" logic.  At current this is neither injected, nor any different to what existed before. The goal is that the "select" logic can do more than look at just the 'folder name' and see if it is in the list of target folders.

For example, it can now be expanded to return true if mutliple things are true, e.g. if the parent folder contains a *.csproj.